sửa file config

+++ src/main/java/com/example/internhub_be/service/DashboardServiceImpl.java
@@ -107,7 +107,7 @@ public class DashboardServiceImpl implements DashboardService {
                     .internId(userId)
                     .fullName(profile.getUser().getName())
                     .positionName(profile.getPosition() != null ? profile.getPosition().getName() : "N/A")
-                    .gpa(evaluation != null && evaluation.getGrade() != null ? evaluation.getGrade() : 0.0)
+                    .gpa(evaluation != null && evaluation.getGrade() != null ? evaluation.getGrade().doubleValue() : 0.0)
                     .status(evaluation != null ? evaluation.getStatus().name() : "IN_PROGRESS")
                     .completionRate(Math.round(completionRate * 100.0) / 100.0)
                     .build();
